feat: placeable mechanic
- Making placeable type item some sort of one way platform
- Making a fall from platform mechanic using s or arrow key down

+++ Assets/Scripts/Items/ItemData.cs
@@ -12,8 +12,8 @@ public class ItemData : ScriptableObject
 {
     public string itemName;
     public int itemID;
-    public int itemWeight;
+    public float itemWeight;
     public Sprite icon;
     public ItemType itemType;
-    public int sellValue;
+    public float sellValue;
 }
